requirements.txt
streamlit
pandas
matplotlib
openai
python-dotenv
xlsxwriter
